requirements.txt
cffi==1.17.1
cryptography==45.0.4
joblib==1.5.1
numpy==2.3.0
pandas==2.3.0
pefile==2024.8.26
psutil==7.0.0
pyAesCrypt==6.1.1
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
watchdog==6.0.0
yara-python==4.5.4
bcrypt
